requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
api==0.0.7
app==0.0.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
init-data-py==0.2.2
magic-filter==1.0.12
multidict==6.0.5
nose==1.3.7
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
ruff==0.6.3
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
yarl==1.9.7
